Moving non-connector specific operations to the api.
