+++ src/api/api.ts
@@ -1,6 +1,7 @@
 import {
   ApiError,
   DeviceData,
+  FileData,
   WorkspaceData,
 } from '@blinkk/editor/dist/src/editor/api';
 import {ConnectorStorage} from '../storage/storage';
@@ -10,10 +11,25 @@ export interface ApiComponent {
   apiRouter: express.Router;
   storage: ConnectorStorage;
 
+  copyFile(
+    expressRequest: express.Request,
+    request: CopyFileRequest
+  ): Promise<FileData>;
+
+  createFile(
+    expressRequest: express.Request,
+    request: CreateFileRequest
+  ): Promise<FileData>;
+
   createWorkspace(
     expressRequest: express.Request,
     request: CreateWorkspaceRequest
   ): Promise<WorkspaceData>;
+
+  deleteFile(
+    expressRequest: express.Request,
+    request: DeleteFileRequest
+  ): Promise<void>;
 
   getDevices(
     expressRequest: express.Request,
@@ -21,9 +37,23 @@ export interface ApiComponent {
   ): Promise<Array<DeviceData>>;
 }
 
+export interface CopyFileRequest {
+  originalPath: string;
+  path: string;
+}
+
+export interface CreateFileRequest {
+  path: string;
+  content?: string;
+}
+
 export interface CreateWorkspaceRequest {
   base: WorkspaceData;
   workspace: string;
+}
+
+export interface DeleteFileRequest {
+  file: FileData;
 }
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
